【fix bug SEQUOIADBMAINSTREAM-74 】节点启动先后顺序会导致LSN较大节点无法当选主
git-svn-id: http://192.168.20.11/sequoiadb/trunk@10692 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/cls/clsVSSecondary.cpp
@@ -108,8 +108,8 @@ namespace engine
          g_startShiftTime = -1 ; // recieve all node sharing-beat
       }
 
-      if ( g_startShiftTime <= _timeout() &&
-           CLS_VOTE_CS_TIME <= _timeout() )
+      if ( g_startShiftTime <= (INT32)_timeout() &&
+           CLS_VOTE_CS_TIME <= _timeout() ) )
       {
          g_startShiftTime = -1 ;
          next = CLS_ELECTION_STATUS_VOTE ;
